Use a simple gap buffer for Buffers.
The original sam implementation used a large number of multilevel
caches, because it had to. When it was written, the operating system's
block cache and virtual memory implementations couldn't be trusted to be
very good...but now, the systems that are targeted by this version of
sam universally have excellent VM and block cache implementations.

Doing this removes a huge amount of complicated code, ends up being
faster in every test I've tried, and removes a lot of the endian-aware
code, meaning that's one less place we have to worry about 64-bitness.

This also paves the way for switching to an encoding-agnostic sam.